Fix cache review findings and CI gates
