+++ crates/rginx-http/src/cache/manager.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod control;
+
 impl CacheManager {
     pub(crate) fn from_config_with_notifier(
         config: &ConfigSnapshot,
@@ -152,7 +154,7 @@ impl CacheManager {
                         base_key: cached_entry.base_key.clone(),
                         key: key.clone(),
                         cache_status: CacheStatus::Updating,
-                        store_response: request.method == Method::GET,
+                        store_response: true,
                         _fill_guard: Some(fill_guard),
                         cached_entry: Some(cached_entry.clone()),
                         cached_metadata,
@@ -252,55 +254,5 @@ impl CacheManager {
         response: HttpResponse,
     ) -> std::result::Result<HttpResponse, CacheStoreError> {
         refresh_not_modified_response(context, response).await
-    }
-
-    pub(crate) fn snapshot(&self) -> Vec<CacheZoneRuntimeSnapshot> {
-        let mut snapshots = self.zones.values().map(|zone| zone.snapshot()).collect::<Vec<_>>();
-        snapshots.sort_by(|left, right| left.zone_name.cmp(&right.zone_name));
-        snapshots
-    }
-
-    pub(crate) async fn cleanup_inactive_entries(&self) {
-        for zone in self.zones.values() {
-            cleanup_inactive_entries_in_zone(zone).await;
-        }
-    }
-
-    pub(crate) async fn purge_zone(
-        &self,
-        zone_name: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::All).await)
-    }
-
-    pub(crate) async fn purge_key(
-        &self,
-        zone_name: &str,
-        key: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::Exact(key.to_string())).await)
-    }
-
-    pub(crate) async fn purge_prefix(
-        &self,
-        zone_name: &str,
-        prefix: &str,
-    ) -> std::result::Result<CachePurgeResult, String> {
-        let zone = self
-            .zones
-            .get(zone_name)
-            .cloned()
-            .ok_or_else(|| format!("unknown cache zone `{zone_name}`"))?;
-        Ok(purge_zone_entries(zone, PurgeSelector::Prefix(prefix.to_string())).await)
     }
 }
