+++ crates/rginx-http/src/cache/runtime.rs
@@ -1,5 +1,11 @@
 use super::*;
 
+mod support;
+
+pub(in crate::cache) use support::PurgeSelector;
+pub(in crate::cache) use support::{build_conditional_headers, remove_cache_files_if_unindexed};
+use support::{stale_if_error_window_open, use_stale_matches_status};
+
 impl CacheRequest {
     pub(crate) fn from_request(request: &Request<HttpBody>) -> Self {
         Self {
@@ -146,10 +152,10 @@ impl CacheZoneRuntime {
     ) -> FillLockDecision {
         let mut fill_locks =
             self.fill_locks.lock().unwrap_or_else(|poisoned| poisoned.into_inner());
-        if let Some(lock) = fill_locks.get(key).cloned() {
-            if now.saturating_sub(lock.acquired_at_unix_ms) <= lock_age.as_millis() as u64 {
-                return FillLockDecision::Wait { waiter: lock.notify.notified_owned() };
-            }
+        if let Some(lock) = fill_locks.get(key).cloned()
+            && now.saturating_sub(lock.acquired_at_unix_ms) <= lock_age.as_millis() as u64
+        {
+            return FillLockDecision::Wait { waiter: lock.notify.notified_owned() };
         }
         let notify = Arc::new(Notify::new());
         let generation = self.fill_lock_generation.fetch_add(1, Ordering::Relaxed) + 1;
@@ -245,82 +251,4 @@ impl Drop for CacheFillGuard {
 pub(crate) fn with_cache_status(mut response: HttpResponse, status: CacheStatus) -> HttpResponse {
     response.headers_mut().insert(CACHE_STATUS_HEADER, status.as_header_value());
     response
-}
-
-pub(super) async fn remove_cache_files_if_unindexed(
-    zone: &CacheZoneRuntime,
-    key: &str,
-    hash: &str,
-) {
-    let _io_guard = zone.io_lock.lock().await;
-    if lock_index(&zone.index).entries.contains_key(key) {
-        return;
-    }
-    let paths = cache_paths(&zone.config.path, hash);
-    let _ = fs::remove_file(paths.metadata).await;
-    let _ = fs::remove_file(paths.body).await;
-}
-
-pub(super) fn build_conditional_headers(headers: &HeaderMap) -> Option<CacheConditionalHeaders> {
-    let if_none_match =
-        header_value(headers, ETAG).and_then(|value| HeaderValue::from_str(&value).ok());
-    let if_modified_since =
-        header_value(headers, LAST_MODIFIED).and_then(|value| HeaderValue::from_str(&value).ok());
-    (if_none_match.is_some() || if_modified_since.is_some())
-        .then_some(CacheConditionalHeaders { if_none_match, if_modified_since })
-}
-
-pub(super) fn matches_vary_headers(request: &CacheRequest, vary: &[CachedVaryHeaderValue]) -> bool {
-    vary.iter().all(|header| {
-        normalized_request_header_values(&request.headers, &header.name) == header.value
-    })
-}
-
-pub(super) fn normalized_request_header_values(
-    headers: &HeaderMap,
-    name: &http::header::HeaderName,
-) -> Option<String> {
-    let mut values =
-        headers.get_all(name).iter().filter_map(|header| header.to_str().ok()).peekable();
-    values.peek()?;
-    let mut rendered = String::new();
-    while let Some(value) = values.next() {
-        rendered.push_str(value.trim());
-        if values.peek().is_some() {
-            rendered.push(',');
-        }
-    }
-    Some(rendered)
-}
-
-fn stale_if_error_window_open(entry: &CacheIndexEntry, now: u64) -> bool {
-    entry.stale_if_error_until_unix_ms.is_some_and(|until| now <= until)
-}
-
-fn use_stale_matches_status(
-    conditions: &[rginx_core::CacheUseStaleCondition],
-    status: StatusCode,
-) -> bool {
-    match status {
-        StatusCode::INTERNAL_SERVER_ERROR => {
-            conditions.contains(&rginx_core::CacheUseStaleCondition::Http500)
-        }
-        StatusCode::BAD_GATEWAY => {
-            conditions.contains(&rginx_core::CacheUseStaleCondition::Http502)
-        }
-        StatusCode::SERVICE_UNAVAILABLE => {
-            conditions.contains(&rginx_core::CacheUseStaleCondition::Http503)
-        }
-        StatusCode::GATEWAY_TIMEOUT => {
-            conditions.contains(&rginx_core::CacheUseStaleCondition::Http504)
-        }
-        _ => false,
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(in crate::cache) enum PurgeSelector {
-    All,
-    Exact(String),
-    Prefix(String),
 }
